Attempted to make this app work as a snap with strict confinement. To do that I changed PacMD to PacTL and etc. However, the app was not able to communicate with pulseaudio. So I had to give up on that for now, and return to this at a later date.
In response, I have changed the confinement level to classic in order
to get the app to work. The compromise there is that the target system
would need to have java installed in order for this app to work.

Added usage, requirements and github url to snapcraft.yaml.

Note to self:
1) You can pull in packages from the repos by using:
	stage-packages
	- name of packages
to a part.