Added tests for all classes and objects.
Fixed formatting in main.kts.

Moved AudioOuput and AudioInput into PacTL,
so that only PacTL can create AudioOuput and AudioInput. This
reduces the chance of invalid data as now, AudioOutput and AudioInput
cannot be created outside PacTL with arbitrary data.

Moved the bash method into PacTL as a private method. Because bash
is only used by PacTL.

Added comments to Main, EventLoop, AudioManager and PacTL to clarify
the purposes of these classes.

Add the Cache class which is utilized by PacTL. This reduces
the load on the CPU at the cost of some extra memory by avoid parsing
the raw data obtained from the pactl commandline tool everytime.
Informal testing has shown a hitrate of around 65%.